+++ core/src/anchor.h
@@ -9,7 +9,7 @@ typedef struct Anchor {
   int last_anchor_col;
   int transpose_state;
   int vertical;
-  float highest_possible_equity;
+  double highest_possible_equity;
 } Anchor;
 
 typedef struct AnchorList {
@@ -21,7 +21,7 @@ AnchorList *create_anchor_list();
 void destroy_anchor_list(AnchorList *al);
 void insert_anchor(AnchorList *al, int row, int col, int last_anchor_col,
                    int transpose_state, int vertical,
-                   float highest_possible_equity);
+                   double highest_possible_equity);
 void reset_anchor_list(AnchorList *al);
 
 #endif
