+++ core/src/leave_map.h
@@ -7,7 +7,7 @@
 #include "rack.h"
 
 typedef struct LeaveMap {
-  float *leave_values;
+  double *leave_values;
   int *letter_base_index_map;
   int current_index;
 } LeaveMap;
@@ -19,7 +19,7 @@ void take_letter_and_update_current_index(LeaveMap *leave_map, Rack *rack,
                                           uint8_t letter);
 void add_letter_and_update_current_index(LeaveMap *leave_map, Rack *rack,
                                          uint8_t letter);
-void set_current_value(LeaveMap *leave_map, float value);
-float get_current_value(LeaveMap *leave_map);
+void set_current_value(LeaveMap *leave_map, double value);
+double get_current_value(LeaveMap *leave_map);
 
 #endif
