+++ core/src/move.c
@@ -93,7 +93,7 @@ void set_spare_move(MoveList *ml, uint8_t strip[], int leftstrip,
            col_start, tiles_played, vertical, move_type);
 }
 
-void insert_spare_move(MoveList *ml, float equity) {
+void insert_spare_move(MoveList *ml, double equity) {
   ml->spare_move->equity = equity;
 
   Move *swap = ml->moves[ml->count];
@@ -108,7 +108,7 @@ void insert_spare_move(MoveList *ml, float equity) {
   }
 }
 
-void insert_spare_move_top_equity(MoveList *ml, float equity) {
+void insert_spare_move_top_equity(MoveList *ml, double equity) {
   if (equity > ml->moves[0]->equity) {
     ml->spare_move->equity = equity;
     Move *swap = ml->moves[0];
